Begin and End function implemented.
Lots still to be done. This is the first step.
SD.begin() can be called with no arguments to use the default ESP32 settings for SD Card connections.

card_connection_mode can be supplied to begin to specify how the SD Card is interfaced with the ESP32 as per the ESP32-EVB or other boards which may use 1  wire mode, SPI or 4 line mode.

Many functions still to be implemented and I am going to rebase these changes onto the top and remove any modifications to SD and FS because any changes made may impact other libraries and I want SD2 to be 'new' and support the default ESP32 functionality.

This is intended to solve the read and write speed and be based off the standard interface provided by arduino.

https://www.arduino.cc/en/Reference/SD